scrupulous/requirements.txt
SQLAlchemy
colander
cornice
pyramid
pyramid_tm
sqlalchemy_utils
transaction
waitress
zope.sqlalchemy

